requirements.txt
Flask>=2.0
python-dotenv
mysql-connector-python
Werkzeug
